requirements.txt
Flask
numpy
tensorflow
opencv-python-headless
Pillow
matplotlib
